# Conflicts:
#	core/src/main/java/info/openrocket/core/database/Databases.java
#	core/src/main/java/info/openrocket/core/file/openrocket/importt/ConfigHandler.java
#	core/src/main/java/info/openrocket/core/material/Material.java
#	swing/src/main/java/info/openrocket/swing/gui/configdialog/ParachuteConfig.java
#	swing/src/main/java/info/openrocket/swing/gui/dialogs/preferences/DesignPreferencesPanel.java
#	swing/src/main/java/info/openrocket/swing/gui/dialogs/preferences/DisplayPreferencesPanel.java
#	swing/src/main/java/info/openrocket/swing/gui/dialogs/preferences/MaterialEditPanel.java
#	swing/src/main/java/info/openrocket/swing/gui/util/GUIUtil.java
#	swing/src/main/java/info/openrocket/swing/gui/widgets/SearchableAndCategorizableComboBox.java

+++ swing/src/main/java/info/openrocket/swing/gui/components/compass/CompassSelector.java
@@ -62,7 +62,7 @@ public class CompassSelector extends CompassPointer {
 		int y = e.getY() - mid;
 		
 		double distance = Math.hypot(x, y);
-		return distance < width / 2;
+		return distance < (double) width / 2;
 	}
 	
 	private void clicked(MouseEvent e) {
