+++ swing/src/main/java/info/openrocket/swing/gui/configdialog/InnerTubeConfig.java
@@ -573,9 +573,9 @@ class ClusterSelectionPanel extends JPanel {
 				double x = points.get(i * 2);
 				double y = points.get(i * 2 + 1);
 
-				double px = BUTTON_SIZE / 2 + x * MOTOR_DIAMETER;
-				double py = BUTTON_SIZE / 2 - y * MOTOR_DIAMETER;
-				circle.setFrameFromCenter(px, py, px + MOTOR_DIAMETER / 2, py + MOTOR_DIAMETER / 2);
+				double px = (double) (double) BUTTON_SIZE / 2 + x * MOTOR_DIAMETER;
+				double py = (double) BUTTON_SIZE / 2 - y * MOTOR_DIAMETER;
+				circle.setFrameFromCenter(px, py, px + (double) MOTOR_DIAMETER / 2, py + (double) MOTOR_DIAMETER / 2);
 
 				g2.setColor(MOTOR_FILL_COLOR);
 				g2.fill(circle);
